Reorganize configuration file structure
